\name{annotationlist}
\alias{annotationlist}
\docType{data}
\title{
 Annotation List Matrix
}
\description{
 Maps the annotation of Affymetrix platforms from the affy [1] package to the annotation names in BioMaRt [2]
}
\usage{data(annotationlist)}
\format{
Matrix 3 rows, 2 columns.
}
\details{
Provided for three Affymetrix platforms HG-U133A, HG-U133A2 and HG-U133Plus2.
}
\source{
}
\references{
[1]Irizarry et al. (2003) Summaries of Affymetrix GeneChip probe level data. Nucleic Acids Research, 31(4); e15.

[2]Durinck et al. (2009). Mapping identifiers for the integration ofgenomic datasets with the R/Bioconductor package biomaRt. Nature Protocols 4, 1184-1191.
}
\examples{
data(annotationlist)

}
\keyword{datasets}
