\name{diseaseRL}
\alias{diseaseRL}
\docType{data}
\title{
Disease Ranked List Matrix
}
\description{
Ranked Lists by expression of genes for reference set of disease profiles.
}
\usage{data(diseaseRL)}
\details{
Matrix containing ranks of genes (rows) for a set of disease profiles (columns). The ranks are in decreasing order of differential expression.
}
\source{
\href{http://www.ncbi.nlm.nih.gov/geo/}{http://www.ncbi.nlm.nih.gov/geo/}
}
\references{
\href{http://www.ncbi.nlm.nih.gov/geo/}{http://www.ncbi.nlm.nih.gov/geo/}
}
\examples{
data(diseaseRL)

}
\keyword{datasets}
