\name{DvDdata-package}
\alias{DvDdata-package}
\alias{DvDdata}
\docType{package}
\title{
DvD Data Package
}
\description{
Data package to provide default ranked profiles, clusters and Cytoscape \href{http://www.cytoscape.org/}{http://www.cytoscape.org/} file format for the DvD package.
}
\details{
\tabular{ll}{
Package: \tab DvDdata\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-16\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
Data loaded automatically in the DvD package.
}
\author{
C. Pacini

Maintainer: Julio Saez-Rodriguez{saezrodriguez@ebi.ac.uk}
}
\references{
Drug Profiles:
\href{http://www.broadinstitute.org/cmap/}{http://www.broadinstitute.org/cmap/}
Disease Profiles:
\href{http://www.ncbi.nlm.nih.gov/geo/}{http://www.ncbi.nlm.nih.gov/geo/}
}

\keyword{ package }
\seealso{

}
\examples{
#load the disease Ranked lists:
data(diseaseRL)
}
