\name{GEOfactorvalues}
\alias{GEOfactorvalues}
\docType{data}
\title{
GEO Factor Values vector
}
\description{
List of possible factor values the GEO data sets can have.
}
\usage{data(GEOfactorvalues)}

\details{
This character vector lists all the available experimental factors which are available in the GEO which can be used explanatory factors in a linear regression model.
}
\source{
\href{http://www.ncbi.nlm.nih.gov/geo/}{http://www.ncbi.nlm.nih.gov/geo/}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(GEOfactorvalues)

}
\keyword{datasets}
